# docs/swagger.yaml
basePath: /
definitions:
  CreateOrderRequest:
    properties:
      product_id:
        example: c0000000-0000-0000-0000-000000000001
        format: uuid
        type: string
      quantity:
        example: 1
        minimum: 1
        type: integer
    required:
    - product_id
    - quantity
    type: object
...
  ErrorDetails:
    properties:
      code:
        example: BAD_REQUEST
        type: string
      details:
        items:
          $ref: '#/definitions/FieldViolation'
        type: array
      message:
        example: Detailed error description
        type: string
    required:
    - code
    - message
    type: object
  ErrorEnvelope:
    properties:
      error:
        $ref: '#/definitions/ErrorDetails'
      requestId:
        example: c56a4180-65aa-42ec-a945-5fd21dec0538
        type: string
      success:
        example: false
        type: boolean
    required:
    - error
    - success
    type: object
  FieldViolation:
    properties:
      field:
        example: quantity
        type: string
      issue:
        example: quantity must be greater than 0
        type: string
    required:
    - field
    - issue
    type: object
...
  OrderResponse:
    properties:
      created_at:
        format: date-time
        type: string
      order_id:
        example: d0000000-0000-0000-0000-000000000001
        format: uuid
        type: string
      product_id:
        example: c0000000-0000-0000-0000-000000000001
        format: uuid
        type: string
      product_model:
        example: MacBook Pro 16 M3
        type: string
      quantity:
        example: 1
        minimum: 1
        type: integer
      remaining_balance:
        example: 500
        format: double
        minimum: 0
        type: number
      status:
        enum:
        - CREATED
        - PROCESSING
        - SHIPPED
        - DELIVERED
        - CANCELLED
        example: CREATED
        type: string
      total_price:
        example: 4998
        format: double
        minimum: 0
        type: number
      unit_price:
        example: 2499
        format: double
        minimum: 0
        type: number
      user_id:
        example: b0000000-0000-0000-0000-000000000002
        format: uuid
        type: string
    required:
    - created_at
    - order_id
    - product_id
    - product_model
    - quantity
    - status
    - total_price
    - unit_price
    type: object
  PaginationMetadata:
    properties:
      page:
        example: 1
        minimum: 1
        type: integer
      page_size:
        example: 20
        maximum: 100
        minimum: 1
        type: integer
      total_count:
        example: 45
        minimum: 0
        type: integer
      total_pages:
        example: 3
        minimum: 0
        type: integer
    required:
    - page
    - page_size
    - total_count
    - total_pages
    type: object
  Product:
    properties:
      category:
        example: laptop
        type: string
      created_at:
        format: date-time
        type: string
      id:
        example: c0000000-0000-0000-0000-000000000001
        format: uuid
        type: string
      manufacturer:
        example: Apple
        type: string
      model:
        example: MacBook Pro 16 M3
        type: string
      price:
        example: 2499
        format: double
        minimum: 0
        type: number
      stock_quantity:
        example: 15
        minimum: 0
        type: integer
      updated_at:
        format: date-time
        type: string
    required:
    - category
    - created_at
    - id
    - manufacturer
    - model
    - price
    - stock_quantity
    - updated_at
    type: object
...
  SuccessEnvelope:
    properties:
      data: {}
      page:
        example: 1
        type: integer
      page_size:
        example: 20
        type: integer
      pagination:
        $ref: '#/definitions/PaginationMetadata'
      requestId:
        example: c56a4180-65aa-42ec-a945-5fd21dec0538
        type: string
      success:
        example: true
        type: boolean
      total_count:
        example: 45
        type: integer
      total_pages:
        example: 3
        type: integer
    required:
    - data
    - success
    type: object
...
  UserSummary:
    properties:
      access_level:
        enum:
        - ALL
        - FILTERED
        - NONE
        example: FILTERED
        type: string
      allowed_categories:
        example:
        - laptop
        items:
          type: string
        type: array
      allowed_manufacturers:
        example:
        - Dell
        items:
          type: string
        type: array
      balance:
        example: 1000
        format: double
        minimum: 0
        type: number
      catalog_access_enabled:
        example: true
        type: boolean
      id:
        example: b0000000-0000-0000-0000-000000000002
        format: uuid
        type: string
      role:
        enum:
        - admin
        - user
        example: user
        type: string
      username:
        example: userA
        minLength: 1
        type: string
    required:
    - access_level
    - allowed_categories
    - allowed_manufacturers
    - balance
    - id
    - role
    - username
    type: object
host: localhost:8080
info:
  contact: {}
  description: Clean, lightweight, and idiomatic Go backend for warehouse inventory
    and transactional ordering.
  title: Warehouse REST API Testbench
  version: "1.0"
paths:
...
  /api/user/orders:
    get:
      description: Returns all orders placed by the authenticated user
      operationId: listUserOrders
      produces:
      - application/json
      responses:
        "200":
          description: List of orders
          schema:
            allOf:
            - $ref: '#/definitions/SuccessEnvelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/OrderResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "503":
          description: Service unavailable
          schema:
            $ref: '#/definitions/ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List customer purchase history
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Atomically purchases a product item, decrements stock, and deducts
        user balance
      operationId: createOrder
      parameters:
      - description: Unique idempotency key to prevent double processing
        in: header
        name: Idempotency-Key
        type: string
      - description: Purchase order request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Order placed successfully
          schema:
            allOf:
            - $ref: '#/definitions/SuccessEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/OrderResponse'
              type: object
        "400":
          description: Malformed JSON syntax or schema validation failure
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "404":
          description: Product or user not found
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "409":
          description: Idempotency conflict or concurrent request in flight
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "422":
          description: Domain business rule violation (INSUFFICIENT_FUNDS, INSUFFICIENT_STOCK,
            FILTER_RESTRICTION)
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "503":
          description: Service unavailable
          schema:
            $ref: '#/definitions/ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Purchase product from warehouse
      tags:
      - User
  /api/user/orders/{id}:
    get:
      description: Returns details for a specific order placed by the authenticated
        user
      operationId: getUserOrderById
      parameters:
      - description: Order ID (UUID)
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order details
          schema:
            allOf:
            - $ref: '#/definitions/SuccessEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/OrderResponse'
              type: object
        "400":
          description: Invalid order ID format
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "404":
          description: Order not found
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "503":
          description: Service unavailable
          schema:
            $ref: '#/definitions/ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Get customer order details
      tags:
      - User
  /api/user/products:
    get:
      description: Browse warehouse items with optional category/manufacturer filters,
        sorting, and pagination
      operationId: listUserProducts
      parameters:
      - description: Filter by category (e.g. laptop, smartphone)
        in: query
        name: category
        type: string
      - description: Filter by manufacturer (e.g. Apple, Dell)
        in: query
        name: manufacturer
        type: string
      - description: Sort field (price, created_at, model)
        enum:
        - price
        - created_at
        - model
        in: query
        name: sort_by
        type: string
      - description: Sort order (asc, desc)
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - default: 1
        description: 'Page number (default: 1, min: 1)'
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: 'Page size (default: 20, min: 1, max: 100)'
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of allowed products with pagination
          schema:
            allOf:
            - $ref: '#/definitions/SuccessEnvelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/Product'
                  type: array
              type: object
        "400":
          description: Invalid pagination, sorting, or filter parameters
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "503":
          description: Service unavailable
          schema:
            $ref: '#/definitions/ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Browse warehouse catalog
      tags:
      - User
  /api/user/profile:
    get:
      description: Returns the profile, current balance, and assigned filter permissions
        for the logged in user
      operationId: getUserProfile
      produces:
      - application/json
      responses:
        "200":
          description: User profile
          schema:
            allOf:
            - $ref: '#/definitions/SuccessEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/UserSummary'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorEnvelope'
        "503":
          description: Service unavailable
          schema:
            $ref: '#/definitions/ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Get authenticated user profile
      tags:
      - User
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      ey..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"